Upgraded webpack, added Typescript for the server, switched UI dependencies from react+babel to svelte. Added some basic docker configuration (just dev so far, prod to come later).
This is part of a complete overhaul of Overseer, starting here with the technology. It previously used multiple sqlite3 databases, now it runs in docker compose alongside postgres and redis (redis is just used for sessions currently).

Script files now generate with a content hash in the name and scripts are cached with a long cache time.