Server rewrite to Typescript
The TS rewrite was fairly involved. First all of the UI/server/shared files were separated (the individual sub folders of each part of the app were all in the same parent folder), then the Stockpile database layer needed to be removed. It's too unmaintainable to keep Stockpile around when I could just use an ORM with TS support, so I switched all of the database access code to use Prisma.

Additionally I used the DB rewrite as an opportunity to switch from sqlite3 to postgres, which also means all of the Booker permissions databases contain all of the permissions, instead of just the permissions for the user in that app. Permissions are also stored in a jsonb column instead of a bitmasked integer. The bitmasking was awkward and didn't gracefully handle adding/removing permissions.

The config.json file has been removed in favor of .env, and keys have been renamed to be more sensible.

I removed a couple things. This is likely no longer going to be hosted internally in my network so the 'proxy' functionality won't be usable, so I'm removing it. Additionally the Hue lights functionality is better handled by Overseer Broadcast and Overseer was using an ancient version of that dependency anyway.

The communication with the Echo server has been greatly improved.
* An integration token JWT for echo needs to be generated in a new integrations page on the admin page and provided on the echo server. Without a token with the proper scope anyone trying to connect as Echo will have their websocket connection refused.
* Echo now relies on the UUIDs (all IDs in the database tables are UUID instead of autoincrement integers) for the file name for every upload instead of keeping the user's original filename. This lets uploads and updates be done without needing to be overly familiar with file names (you needed to stay away from uploading generic file names like 'setup.zip' because that would turn into the unique identifier for the game, every update needed the same file name to overwrite it), and now because the file name has nothing to do with the item in Echo it matches you need to go specifically to the edit page for something in Echo to upload a new file.
* Uploading something has a bit more process to it. Previously it'd send metadata to Overseer and post the file to Echo, now when metadata is sent to Overseer it returns a URL to post the file to, this URL includes the new ID for this item. The ID is also shared to Echo before sending it back to the browser as Echo will refuse every upload it's not expecting.
* There's no more awkward merging/creating of echo metadata ("new" and "update" are separate actions now). It was a weird solution for me being lazy when migrating things, I hadn't wanted to reupload everything just to specify the new metadata, but now that everything needs a new ID that'll be required anyway. Because of this the Echo server cannot implicitly create items in Echo just because the file exists in its storage directory. It needs to be created in Overseer first.
* Downloads need to provide an integration token, this is an automatically generated JWT given to users who have the 'download' permission. They can pass this token along when downloading anything from Echo, Echo will verify it with Overseer before serving the file.
* More care is taken to validate (using Ajv) metadata changes.

"Links" as a separate concept from notes for candidates in voter was an awkward pattern to manage for the UI. Instead Voter and Echo now support markdown in their respective 'notes' sections so you can include whatever you want there that's supported by markdown. Removing this removed a lot of specific code in Voter for link CRUD operations and suggesting a link title (notes are free form so you're on your own for link text).

A general refactoring pass was done over everything. Not everything is done and tested, but this is the majority of the changes to the server that are expected.