The admin page has been rewritten with svelte/sheodox-ui!
Roles/Assignments/Integrations(new) are now on their own tabs instead of the page being a crazy dump of everything related to admins. It functions basically the same way with the exception of the new Integrations tab which lets you generate an integration token for Echo (and in the future maybe other things).