Rewrote the front end in Svelte.
Everything was a big React app before. Now (most of) it has been rewritten in svelte. Voter sort locking has yet to be done, and images on candidates aren't supported yet.

Voter and Echo have notes done in markdown, and the view/edit UIs are separate so when viewing something you don't just get the same textarea you edit it in anymore.

Echo has been rewritten internally, as a result updates aren't done by uploading a new zip with the same file name (the original file names are discarded now) and instead you need to go to the "edit" page on an item. Edit/Upload now share the same component (previously view/edit shared a component, but with the markdown changes a much better dedicated view component could be built). From a UI perspective it's fairly similar though, but with a lot less unnecessary space wasting (tags are only shown when toggled on when viewing the list, and only the applied tags are visible when viewing something, not the whole cloud). Tags when viewing an item in Echo are now links to a search for that tag.

Voter has had lots of changes, though it's not yet done. Instead of selecting a race with a dropdown you're presented with a dashboard of all races (and their top three candidates) from there you can follow a link directly to the race. Again less wasted space!

Currently I haven't rewritten the switchboard. I'll probably make it show some relevant previews from the different apps similar to the Voter dashboard. The LoginRequired component also needs a bit of work. The header now has links to echo/voter so the switchboard itself isn't as useful.

With svelte's component based styling and the theme provided by sheodox-ui I've not needed custom styles yet so all of the old theming scss has been deleted (except admin.scss).

The old three spinning triangle loading indicator that was done with Threejs/GLSL has been rewritten to be just CSS/HTML so it's a bit more lightweight and multipurpose.