Re-added sort locking.
When you're voting it's disorienting if candidates keep moving around as you and other people vote, so when the list is being hovered over it will no longer sort immediately. Changes will be merged so your voting status is reflected and you can see new/deleted candidates, but as soon as the lock releases (mouse leaves the list) it'll sort regularly.

Deleted candidates turn grayscale and voting gets disabled, expanding still works but the options button gets removed.

Added a key for the user bubbles on CandidateVoteBar.svelte. I don't know why exactly but it would occasionally get a null somewhere in there and throw an error. Adding a key seems to fix it. It is however using the whole user object as a key because IDs aren't provided on masked users. It will probably rebuild all of the user bubbles but that probably shouldn't matter that much.