Added images to Voter candidates.
Images can be uploaded by pasting into the Notes textarea when editing a candidate, thumbnails are viewable in candidate details when they're expanded. Clicking one of those thumbnails will show a big version of that image, with horizontally scrolling thumbnails below it (clicking one switches to showing that one big). It works like you'd expect from other sites. On the candidate editing modal you'll see a list of those same images with a delete button, this is where you're able to remove them. Like with everything else out there I wanted to keep editing and viewing separate so unlike with the react version of the site you can't delete images while viewing them. I think hiding the destructive/edit options away in general makes using the site a lot more friendly for people just browsing which is what most people will do.

Because I needed a place to show the images you could delete I decided it was better to make all of the editable stuff editable within one modal, it's a lot easier to edit things in one place and matches how it's done in Echo.

I had to add a 'source' column to the image table so the right permissions can be checked when serving an image. Images were previously checked against the source specified in the route, but that left it vulnerable to requesting images that belong to something the user doesn't have access to and don't actually match the source they specified. For example, when echo has images and a user has access to view echo but not voter, if they knew a UUID for a voter image they could have requested `/images/echo/uuid-of-something-in-voter` to view a voter image using the permissions check against echo. To get around this the source is no longer required in the path, instead the source is a required column in the database, and the source is mapped back to a voter for which to check a 'view' permission.

The nginx config also needed a bump in the max body size, otherwise you couldn't upload high res images. High res images aren't a problem because they're resized anyway.