Logging and error handling.
Using a few different loggers that log with a 'concern' property based on what logger is used. This makes it easy to determine what part of the app the logs came from. Most of the usage so far is just through some error handlers for websockets and express routes. I'm using a strategy similar to what I did with Context.Reviews.

Expose an internal server running on port 4001 (still proxied through nginx on that port) that exposes an endpoint for logs. The endpoint requires a 'logs' scoped integration token.

The websocket safety handler lets you give it an action, it'll check to make sure you have that permission and send the user a toast if they don't have permissions to do something. If an error occurs when handling the websocket message it'll log it like the express middleware.

Booker now clears its permissions cache when assigning roles. It wasn't previously doing so so you couldn't give permissions without restarting the server or touching a role's configuration.

All requests get a requestId which is logged with handling errors from the error handler middleware (and given to the user).