When the form is submitted, disable the submit button to prevent double uploads!
Also moved the upload button and image previews to be in the right column underneath notes instead of beneath both columns. Previously the more tags there were in the tag cloud the more the images and upload button would be pushed down, now that they're in the notes column that won't happen!