Added an "Events" planning feature.
Events uses three tables. An "Events" table for every actual event that's planned, an "RSVP" table for every response to an event, and an "RSVPDay" table that has day-by-day responses to every day an event is held, so a user can RSVP to an event but only plan on attending certain days if the event lasts more than one day.

With events I used a single 'organize' permission instead of separate add/update/delete permissions. Realistically I can't see a situation where someone can update but not delete an item etc