Remove of the user specific app data masking.
For some reason I had gone through a lot of effort to hide user IDs from everyone, with the exception of your own nowadays in the `window.user` serialized bundle, but it's not very efficient or necessary. The user IDs are auto generated uuids so no private information can be gleaned from them. I'm taking a step back and sending user IDs as is. Whenever there was a user ID somewhere Overseer would "mask" it and replace it with the displayName and profileImage for that user, but that is often repeated a lot of times, leading to quite a bit of redundant data being sent.

Now user IDs are sent as is everywhere (except on the admin page where more information is needed) and the UserBubble component will try and load any users it hasn't yet loaded by requesting that "masked user" data if it hasn't yet done so for that user ID. A new generalized "appBooker" was made for this general stuff.

A lot of user specific stuff was done when preparing event/voter data for the client to identify that user's relationship to the data (how they voted, if they created a candidate etc, how they RSVP'd to an event) but that was able to be taken out now that the front end can calculate that when it gets the data.