Efficient voter updates.
Whenever anything happens in voter it would send the entire dump of the database basically. Even for a single vote it'd do that, with many people connected and many votes going out the duplicate data gets ridiculous. I'm instead using `deep-diff` to send the full data on init only, and every broadcast only sends minimal diffs so the changes can be reconstructed on the front end by each client.

I realized I was sending a redundant 'votes' array, I've removed it now because I wasn't using it on the front end and it'd balloon the diffs quite a bit.