Start of notifications!
Notifications currently are basically toasts that get persisted. Not all toasts will make it to a full persisted notification (error toasts for example) but important things like uploads to Echo, and new events will be. This like new candidates in voter are only toasts, because they're not so important that you need to care about them unless you're actually using the site at the time.

Notifications belong to individual users, so if something happens a new notification row is stored for every user that has access to it and each has its own "read" boolean.