Partially rewrote "Conduit/SilverConduit". I renamed them to "Envoy" and "Harbinger" respectively because I didn't like the "Conduit" name. They're functionally the same as before, but now Harbinger doesn't require 'io' to be passed (it can import it by itself) and it isn't a subclass of Conduit, plus it can broadcast to all connected clients of an individual user. This is necessary for user-specific things like toasts, but things we want to keep in sync between every client for that user, like toasts that have been read, or new toasts. It's similar to the filteredBroadcast but without requiring the consumer of the method to specify a filter function- you just give it the event, userId, and data.
Envoy now also doesn't have a concept of "subs". THey were meant to be an allowance for nested handlers, but in practice I never actually used them.

Also made it log errors when handling a notification route to the console directly during development so the error stack trace and message don't get mutilated and put all on one line in the JSON logging.