Added push notifications and event reminders.
This requires a service worker, which also meant HTTPS is needed for development so the nginx config was tweaked.

Push notifications require permission to be granted in the browser, after which we'll get a payload that specifies the endpoint to send notifications as well as encryption/authentication details. The endpoints are device specific so a user may have multiple push subscriptions so multiple rows with the same userId are allowed in the UserPushSubscription table, if a status code 410 response is received when sending a notification it means the endpoint is no longer valid so that subscription gets deleted from the database. I'm not sure if I'll hit any others, for now any other status code just logs the error. I don't want an internet blip or something unexpected to wipe out all of the subscriptions! The actual pushing is handled by the web-push package.

Clicking on a notification will redirect the user to a `/n/:notificationId` endpoint which will mark the notification as read before redirecting to whatever the href of the notification was. This is done so it doesn't present the notification as if it's unseen, because clearly you saw it if you clicked it.

I wanted notifications to be enabled or disabled by type so I had to add a settings menu (which required a user dropdown in the header because I ran out of space) where you're presented with settings for all that and another always present way to give permission for push notifications (on the event page it'll prompt you to give permissions but you can click "no, don't ask again" so there needed to be another way to turn them on if you already clicked that).

Every 30 seconds Overseer now queries for events that start within the next 24 hours and check if any of them haven't sent out a reminder yet. If it's within 24 hours it'll send a reminder out saying the event starts in a day, and if it's within an hour it'll say that too. Notifications go out to everyone who is able to view events and hasn't RSVP'd as not going, so any likely or possible attendees. It'll also encourage them to RSVP if they haven't done so. Care was also made to not remind people about events that start too soon. If an even starts less than two days from the time of creation it won't send a one day reminder, same for two hours and the one hour reminder. Without those considerations you'd send conceptually redundant notifications if the event is made on short notice.

Push notifications need to be encrypted, so Overseer now automatically generates and stores (in a new AppMetadata table) some VAPID public/private keys for them.

A new tab on the admin page was added for creating site wide announcements. This can be used for things like notifying of event time changes or whatever else.

I added some shortcut allowAll/denyAll methods for booker roles, this lets you quickly make a role on the admin page with full permissions instead of checking a ton of boxes. Good for when I need to reset the database during development.