Bootstrap the application with data on page load
This should speed up initial page load, as well as prevent the awkward jumping of data on the home page. When the data is loading and the homepage is viewed all three columns for Events/Echo/Voter show up, but if there are no relevant events to show the Events column disappears. The page jumping around isn't very good looking.

This change embeds the initial state of all three apps in the script tag where the user data and other things are serialized. When the stores are initialized they'll check for data in the __INITIAL_STATE__ global, and the stores won't try to fetch data over the websocket. This means that as soon as the page loads all data is shown already. This has an added benefit of being more efficient with data, as the page html is gzipped but the websocket isn't.

A little bit of refactoring needed to be done for this. The router files all needed to export a function that returns their cached data (except for events because it's a little different with the permissions), and to export those functions they needed to be using the import/export syntax so I needed to mess around with imports. I also made the functions to fetch the data for that part of the application follow better naming conventions (echo's `prepareData` function is now either `getEchoData` or `getNewestEchoData` depending on if you are ok with the cached version.