Improved event view layout.
When viewing an event the UI was kind of crowded, especially as an organizer.

For organizers I moved the options menu (a menu button with just a chevron-down icon to be below the event details. It wasn't really necessary to have them be at the top, and it'd get in the way because that row is already crowded. I added an "Event Options" button text so it's more descriptive.

The metadata (virtual/in-person, and the time) is now smaller and has a muted text color. It's important but the extra font size in the area looked bad and I'm using the muted color to differentiate it from the event details, and the `.gap-4` on them meant that when they wrap there was awkward vertical space. That's now just a right margin on the attendance type.

The header row can now wrap, so instead of the header and rsvp buttons each awkwardly individually wrapping to maintain the three columns in one row the RSVP button can go onto its own row.