Create an echo download token and pass it along with the initial state.
Emitting init on the echo websocket causes echo to respond with the data, and a download token if you have permissions. Because we now skip the init message with the data being bootstrapped to the page we lost that. This creates a download token and sends it along with the rest of the echo data.