Better handling of push subscriptions
This ensures when the page loads that Overseer knows of the most recent subscription. Sometimes the browser vendor (Firefox at least) will expire old subscriptions and replace them, and if that's happened we won't know of the subscription anymore but the browser will think it's subscribed. This should hopefully fix a bunch of currently broken subscriptions.

I also added some serviceworker code that should hopefully notify overseer of any changes with a pushsubscriptionchange event. I can't figure out how to test it so I'm just going to hope for now.