Build the frontend assets while the image builds
The frontend assets get compiled and put in a folder that's shared with nginx. This was previously done on boot because the folder is a bind mount and only is mounted at runtime. This change builds the assets to an intermediate folder which then gets copied to the public folder on boot, copying a folder is much quicker than the build so this should result in significantly less downtime deploys.