Front end typescript rewrite!
This includes a new version of sheodox-ui which is written in typescript. Typings were added to src/shared/types, and used on both the frontend and backend. I also switched to esm and because of how esm/typescript (currently?) evaluates import paths the server needs to include `.js` in every file name.

I also switched to using Vite to build front end scripts/css, which has hot module replacement build in. The build scripts for the production image aren't yet finished, but that should be easy. I'll probably still use webpack for non-script things for now (copyplugin).