Fix prod build and static asset copying, remove webpack
Because webpack was replaced with vite I needed a new way to copy some static assets to the public folder. Vite can handle this with a publicDir configuration option, but it doesn't actually copy the files in dev, which didn't work for the way this project is set up. Additionally it can't copy files from other directories so I couldn't copy fontawesome files that way. Instead I just wrote a tiny bash script that copies the necessary folders over before the build and before building/running the container (to ensure the folder owner is the user etc). I also changed the file paths of the static assets a bit to have a simpler structure.