Minor tweaks to Echo styling
This moves the upload button to be within the same column as the echo item details instead of in a row above all of the fields.

This also removes some spacing between the image and the description in grid view, and fixes the border radius used on multiple elements so the whole things has a contiguous border with rounded corners, without making the border on each element get cut off (like the blue border on hover).