Name the session ID cookie, and fix an issue where a user who isn't logged in was shown a "no permissions" screen instead of a login link.
There were some template conditionals that looked for the existence of a user object, but the user object now always exists, but some properties on it might not exist (for a consistent interface with UserBubble with the user registry). For now just checking that the user has an ID is enough.