Mobile styling tweaks.
Just a bunch of tweaks to the width of things, with specific care for mobile screen sizes.

This also reorders the home page so Voter shows before Echo. So if a user needs to vote on something while on their phone they don't have to scroll past Echo, which is pretty useless on a phone.