Use tabs to show all of the event rspv information
Showing all the RSVP information at once was kind of tedious and it never really looked right.  This change moves the RSVPs, Notes (if you're an organizer), and RSVPs by Day and Overnight Attendees (if the event is multiple days), to separate tabs. This looks so much better than the layout looked previously.