Reorganized the wizard voting buttons
The vote buttons were too far apart, you'd have to mouse over the Back/Skip buttons if you wanted to vote the other direction than the one you last did. Back/Skip are now in the same row as the -/+ are, the voting buttons are in the center too. This should make it less cumbersome to vote.