Album and wizard style changes
Increased the z-index of the theater so it shows up over the wizard, as the theater can be launched from the wizard it needed to be on a higher layer or it opens in the background.

The album image sizing has been tweaked, for a strip the image now has a max width of 90% instead of 90vw. The 90vw size was causing the images on the vote wizard to always be too wide on mobile causing unwanted horizontal scrolling. The 90vw is still needed for cover, otherwise the images are smaller than the card they're on on Echo.