Shuffle candidates on the wizard
I wanted users to not get accustomed to the order they see things in when they open the wizard so now it will randomize the order of the candidates. The one extra condition is that it orders all of the candidates they haven't voted on first, then all the ones they have voted on, just within those two groups they're in a random order.

I also tweaked the vote button styles to now indicate what the user's current vote is for a candidate using styles similar to the candidate list itself.