Ability to filter ranking to specific voters
If we're deciding which game to play and a lot of people have voted but only a few of them are here it's difficult to figure out what won in a race between only the people who are available. This adds the ability to filter out specific voters, so if they're not present they don't get a say in what game is played by those who are.

I added a new "Show Filters" menu item on Voter. This will show a bunch of buttons for everyone who has a vote in the current race, implicitly everyone's vote is counted and clicking on a user will filter out their votes. This was easier than the inverse of keeping a list of allowed voters (you'd have to merge in new voters as they appeared and keep a list of voters that aren't included anyway to know who to auto-add to the list of allowed voters in that case).

The ranking now takes this filter list into consideration.