Candidate banning
Some items in a race aren't really viable depending on the situation, like a game night with ten people shouldn't allow voting on four player campaign games. This adds the ability for users with the `ban_candidate` permission to "Ban" a candidate, which clears votes and disallows voting. Additionally banned candidates will always show at the bottom of the ranking.