Styling consistency and layout changes
This is mostly a bunch of style changes to make padding/margins/borders/border-radius more consistent particularly on things that show on the home page.

Headings on most pages are bigger now. Some unnecessary text was cut from voter previews.

Updated sheodox-ui which now has a centered nav slot in the header so the app within Overseer you're on is more obvious.

The home page now shows just two columns, if there's an event to show it shows as a new row above Voter/Echo, as there would only ever be one event that shows it seemed unbalanced.