Update candidates with your voting state while voting on the wizard
It seems when voting the candidates stored in the voting wizard don't get updated when the user votes one way or another. They're copied from an array of candidates when the wizard opens to avoid it changing with other people's votes, but that also means it doesn't seem to get updated as you vote. So voting for something then hitting back doesn't show you what you just chose. I made it so it'll in-place set the voting state as you're voting for things. This should work most of the time (doesn't accurately reflect your voting state between computers probably, but people probably won't vote on two computers on the same account at once).