Don't automatically redirect to the newly created race
When the new race is created and redirected to we won't yet know the race details so it just throws an error. Redirecting there is nice, but since updates are sent as diffs instead as commands to mutate store like my newer projects we can't be more surgical without having some pretty ugly code with every update. Just removing that instead.