Overseer Echo server rewrite
This required a rewrite of the websocket code to use plain websockets, as the echo service now is written in Go and can't use socket.io.

Things work similarly, but the request-response pattern is a little more manual. A message can send an object with a msgId string, and if something is responding to that message it'll send it back which calls a handler.