Post Echo upload as a file directly, not as a form
The file was being wrapped in a FormData object and posted as a multipart form, but this was causing complications with the Go service that handles the upload. It was parsing the whole body to a temporary file, then parsing the file out of it and saving that, so the entire upload would be written to disk twice. Posting the entire file by itself seems more direct and cuts out the intermediate step.

Also updated sheodox-ui so slow updating (or errored) progress toasts won't expire, leaving users unsure of the progress. Also updated some other dependencies.