Cache bust icons
Using Fontawesome 6 instead of 5, but the icons are cached super long, so I needed to cache bust them, the easiest way is just putting them in a 'fa6' folder so if I do a major upgrade i can just bump that version. I couldn't get it to import other ways but this will be fine.