Update sheodox-ui, more border radius
Going for a more rounded elements, it looks more modern, using that and some extra gap on the event view.

Fixed a bug that caused the event time text end date to be skipped when the start/end are a day apart, it was meant to skip if they're the same day to make it easier to understand at a glance.