Fix mobile view of events, show "No RSVP Yet" for status == ''
This removes the divider between events on the home page because it got in the way when wrapped into a column, and this is easier than switching orientations.

Also made status badges show "No RSVP Yet" in a similar style to the other RSVP events as a nudge to action, and to make the not-rsvp events be just as tall as the rspv'd events on the homepage.