Show candidate details in a modal, more border radius
This replaces the inline expanding panel for candidates with a modal. The intention of expanding all the details was originally so you could make informed voting choices more easily without having to view each candidate's details individually, but that's an outdated concept now that "Help me vote" is a thing.

Also changed border radius and padding of a few things to look more modern.